of/pci: allow some devices on a bus to bypass an iommu
The iommu-map binding allows to assign different devices under a root
complex to different iommus. Nothing prevents a subset of devices to
bypass an iommu entirely and access the physical address space. It is
especially useful for VMMs that want to put virtual devices with different
IOMMU capabilities under the same root complex for simplicity.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>