iommu/of: Allow drivers to defer probe indefinitely
Because its transport code can be built as module, the virtio-iommu driver
may also be a module and will be loaded by userspace. Currently the core
gives up deferring endpoint probe after the initcalls and allows the
driver to bypass the IOMMU layer, which results in translation faults if
there was an uninitialized IOMMU.

Allow IOMMU drivers to be built as modules. If the device tree describes
the endpoint to be managed by a virtio-iommu and the driver is a module,
don't give up probe deferral.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>
---
I'd be tempted to remove the "if IS_MODULE", since we probably don't want
endpoints to bypass the IOMMU anyway.