KVM: arm64: nv: Opportunistically detect HCR_EL2.E2H being flipped
With NV2, we cannot detect writes to HCR_EL2, meaning that we are
at the mercy of the guest flipping HCR_EL2.E2H from one mode to
the other.

There is no good fix for this, but we can at least act on a
transition if we have trapped for any reason. If this was a 0->1
flip and that we have trapped because of an EL1 register access,
we will be able to flip NV1 and retry.

But this is only papering over an architecture problem, and the
correct fix is to make KVM advertise E2H being RES1.

Signed-off-by: Marc Zyngier <maz@kernel.org>