iommu/virtio: Add a new geometry property
The VIRTIO_IOMMU_PROBE_T_GEOMETRY allows a device
to describe which IOVA geometry can be used by the driver.
Together with the reserved regions, this allows the guest
to use valid IOVAs.

If the device exposes this property, use it to set
a geometry for the associated domain. Otherwise use
the geometry exposed in the device config.

Signed-off-by: Eric Auger <eric.auger@redhat.com>